PLUGINS-1287: Adapt TinyMCE5 to new structure
As the plugins for TinyMCE4 and TinyMCE5 share the same source file (editor_plugin.src.js), this file has been symlinked from packages/mathtype-tinymce4 to packages/mathtype-tinymce5.
It is important to keep in mind that the replacements applied to the source in TinyMCE4 will affect TinyMCE5 as well. I hope this doesn't cause much disruption from this point until the variable replacements become obsolete.