PLUGINS-1299: New CKEditor 5 plugin
The CKEditor 5 plugin has been refactored. In theory, this should not be a breaking change so anyone using the plugin need not change any code on their side.
The old plugin used a model mapping that represented all the MathML tags as elements in the model.
This was good for faithfully representing the information, but bad for interacting with our view (the rendered images).
To cope with this, we have changed to a new "flat" model, which consists of a single `<mathml>` element with a `formula` attribute that contains all the MathML as a string.