Fix the file pattern to add react demo on lerna
The packages key on the lerna.json file needs
to include all folders and sub-folders on the
demos folder.

To avoid this issue, we've add a sample lerna.json
file, with the right configuration.

In any case, changing the lerna.json to compile
the demos does not seems like the right solution.

We could include a demo folder on every package?