Update bootstrap process and README instructions
The Bootstrap process of lerna was raising
errors if you followed the instructions on
the Quick Start section of the README file.

The 'demos/ckeditor5/package-lock.json'
needs to be deleted before executing the
'lerna bootstrap' command. Otherwise it
raises an error.

Also, there's the need to update manually
the lerna.json file to compile the Technical
Demos. This was not indicated on the README.

I've simplified the bootstrap process: it
takes npm install & npm start to run all.

The README file has been updated with
the right commands.