Fix multiple issues on real and simulated iOS devices
1/

Fix errors affecting UserAgent simulation scenario.

An error stops execution when opening MathType with some
iOs simulated user agent. The error differs depending on the browser.

On real iOS devices everything works as expected and there are no errors
on the console, though.

The bug is fixed by making sure the 'modalDialogInstance' is available on
execution time.

2/

Load the editor in fullscreen mode on iOs devices.

Fixed by improving the browse detection method for iOs devices
on the contentmanager class.

Also:

- Refrain some instructions for better understanding.

- Make devserver accessible on all html5 demos.

In order to test this on actual mobile devices, one must expose the
demos to the local network. To do this, just modify the
webpack.config.js file for the desired demo so as to include
host: '0.0.0.0'  in the devServer section.

See: https://webpack.js.org/configuration/dev-server/#devserverhost

- Show the editor in modal mode to android/iOs users.

- Fix caret getting inside the <span> in Firefox