Fix errors affecting UserAgent simulation scenario
An error stops execution when opening MathType with some
iOs simulated user agent. The error differs depending on the browser.

On real iOS devices everything works as expected and there are no errors
on the console, though.

The bug is fixed by making sure the 'modalDialogInstance' is available on
execution time.