Freeze Froala 3 version to 3.2.1 on the html5 demo
Even though the latest release 3.2.6-1, from 5th
feb, says it fixes ""Method events.focus() does
not work" the previously identifed error persists.

Therefore, the best solution is to freeze the
Froala3 dependency version to 3.2.1 until no
further test is performed and/or wait for Froala
to fix the issue with 'events.focus()'.

- https://github.com/froala/wysiwyg-editor/releases/tag/v3.2.6-1

Also on this commit, for the Froala3 demo:

- Fix dependency vulnerabilities.
- Fix package.json dependencies: convention.
- Remove 'linting' at demo level.
- Upgrade to lerna 4.0.0 at demo level.
- Remove uglify lib since it was not used.
- Update demo authorship to team level.

#see: KB-10218