Fixed 'Moodle install folder is not root'
> This change is based on 'stable' branch.

The 'integrationModelProperties.configurationService' value
did not included the 'M.cfg.wwwroot' value.

Therefore, the plugin did not load correctly when Moodle was
installed on a different folder than the root. Like:

- http://domain.com/moodle/

Instead of:

- http://domain.com/

The library has been compiled using the 'moodle' comand to
be able to generate a new version to be included as a fix for
the Moodle Integration plugin: moodle-tiny_mce_wiris.

#taskid KB-13266