Use link-checker to validate broken links
## Use link-checker to validate broken links

Use link-checker instead of check-html-links to look for broken links on
the static site generated with the jsdoc command.

This library offers more parameters to exclude which links to validate.

Added a configuration file at the main root folder.

When everything works fine this next message should appear:

"0 errors and 0 warnings"

## Added a general lint command

A general lint command has been added that lints all kind of files at
once.  The command is:

$ npm run lint