doc: add limitation disclaimer about inline limitations
As documented on issue #486.

When using Safari, after inserting a new formula with ChemType
editor, the caret is lost on the editor's target element,
and the focus is lost, too. The caret should be placed after
the new element added. If the formula is being added after
clicking the MathType editor, it works as expected.

Also on this commit:

- Fix the documentation on the TinyMCE5's README.

$taskid: 11529